Refactor index reader and chunk reader
