+++ pkg/storegateway/bucket.go
@@ -38,7 +38,6 @@ import (
 	"github.com/weaveworks/common/httpgrpc"
 	"golang.org/x/exp/slices"
 	"golang.org/x/sync/errgroup"
-	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 
@@ -922,321 +921,6 @@ func (s *BucketStore) limitMaxTime(maxt int64) int64 {
 	}
 
 	return maxt
-}
-
-type seriesEntry struct {
-	lset labels.Labels
-	refs []chunks.ChunkRef
-	chks []storepb.AggrChunk
-}
-
-// blockSeriesClient is a storepb.Store_SeriesClient for a
-// single TSDB block in object storage.
-type blockSeriesClient struct {
-	grpc.ClientStream
-	ctx             context.Context
-	logger          log.Logger
-	extLset         labels.Labels
-	extLsetToRemove map[string]struct{}
-
-	mint           int64
-	maxt           int64
-	indexr         *bucketIndexReader
-	chunkr         *bucketChunkReader
-	loadAggregates []storepb.Aggr
-	chunksLimiter  ChunksLimiter
-	bytesLimiter   BytesLimiter
-
-	lazyExpandedPostingEnabled                    bool
-	lazyExpandedPostingsCount                     prometheus.Counter
-	lazyExpandedPostingSizeBytes                  prometheus.Counter
-	lazyExpandedPostingSeriesOverfetchedSizeBytes prometheus.Counter
-
-	skipChunks             bool
-	shardMatcher           *storepb.ShardMatcher
-	blockMatchers          []*labels.Matcher
-	calculateChunkHash     bool
-	seriesFetchDurationSum *prometheus.HistogramVec
-	chunkFetchDuration     *prometheus.HistogramVec
-	chunkFetchDurationSum  *prometheus.HistogramVec
-	tenant                 string
-
-	// Internal state.
-	i                uint64
-	lazyPostings     *lazyExpandedPostings
-	expandedPostings []storage.SeriesRef
-	chkMetas         []chunks.Meta
-	lset             labels.Labels
-	symbolizedLset   []symbolizedLabel
-	entries          []seriesEntry
-	hasMorePostings  bool
-	batchSize        int
-}
-
-func newBlockSeriesClient(
-	ctx context.Context,
-	logger log.Logger,
-	b *bucketBlock,
-	req *storepb.SeriesRequest,
-	limiter ChunksLimiter,
-	bytesLimiter BytesLimiter,
-	blockMatchers []*labels.Matcher,
-	shardMatcher *storepb.ShardMatcher,
-	calculateChunkHash bool,
-	batchSize int,
-	seriesFetchDurationSum *prometheus.HistogramVec,
-	chunkFetchDuration *prometheus.HistogramVec,
-	chunkFetchDurationSum *prometheus.HistogramVec,
-	extLsetToRemove map[string]struct{},
-	lazyExpandedPostingEnabled bool,
-	lazyExpandedPostingsCount prometheus.Counter,
-	lazyExpandedPostingSizeBytes prometheus.Counter,
-	lazyExpandedPostingSeriesOverfetchedSizeBytes prometheus.Counter,
-	tenant string,
-) *blockSeriesClient {
-	var chunkr *bucketChunkReader
-	if !req.SkipChunks {
-		chunkr = b.chunkReader()
-	}
-
-	extLset := b.extLset
-	if extLsetToRemove != nil {
-		extLset = rmLabels(extLset.Copy(), extLsetToRemove)
-	}
-
-	return &blockSeriesClient{
-		ctx:             ctx,
-		logger:          logger,
-		extLset:         extLset,
-		extLsetToRemove: extLsetToRemove,
-
-		mint:                   req.MinTime,
-		maxt:                   req.MaxTime,
-		indexr:                 b.indexReader(),
-		chunkr:                 chunkr,
-		chunksLimiter:          limiter,
-		bytesLimiter:           bytesLimiter,
-		skipChunks:             req.SkipChunks,
-		seriesFetchDurationSum: seriesFetchDurationSum,
-		chunkFetchDuration:     chunkFetchDuration,
-		chunkFetchDurationSum:  chunkFetchDurationSum,
-
-		lazyExpandedPostingEnabled:                    lazyExpandedPostingEnabled,
-		lazyExpandedPostingsCount:                     lazyExpandedPostingsCount,
-		lazyExpandedPostingSizeBytes:                  lazyExpandedPostingSizeBytes,
-		lazyExpandedPostingSeriesOverfetchedSizeBytes: lazyExpandedPostingSeriesOverfetchedSizeBytes,
-
-		loadAggregates:     req.Aggregates,
-		shardMatcher:       shardMatcher,
-		blockMatchers:      blockMatchers,
-		calculateChunkHash: calculateChunkHash,
-		hasMorePostings:    true,
-		batchSize:          batchSize,
-		tenant:             tenant,
-	}
-}
-
-func (b *blockSeriesClient) Close() {
-	if !b.skipChunks {
-		runutil.CloseWithLogOnErr(b.logger, b.chunkr, "series block")
-	}
-
-	runutil.CloseWithLogOnErr(b.logger, b.indexr, "series block")
-}
-
-func (b *blockSeriesClient) MergeStats(stats *queryStats) *queryStats {
-	stats.merge(b.indexr.stats)
-	if !b.skipChunks {
-		stats.merge(b.chunkr.stats)
-	}
-	return stats
-}
-
-type sortedMatchers []*labels.Matcher
-
-func newSortedMatchers(matchers []*labels.Matcher) sortedMatchers {
-	sort.Slice(matchers, func(i, j int) bool {
-		if matchers[i].Type == matchers[j].Type {
-			if matchers[i].Name == matchers[j].Name {
-				return matchers[i].Value < matchers[j].Value
-			}
-			return matchers[i].Name < matchers[j].Name
-		}
-		return matchers[i].Type < matchers[j].Type
-	})
-
-	return matchers
-}
-
-func (b *blockSeriesClient) ExpandPostings(
-	matchers sortedMatchers,
-	seriesLimiter SeriesLimiter,
-) error {
-	ps, err := b.indexr.ExpandedPostings(b.ctx, matchers, b.bytesLimiter, b.lazyExpandedPostingEnabled, b.lazyExpandedPostingSizeBytes, b.tenant)
-	if err != nil {
-		return errors.Wrap(err, "expanded matching posting")
-	}
-
-	if ps == nil || len(ps.postings) == 0 {
-		b.lazyPostings = emptyLazyPostings
-		return nil
-	}
-	b.lazyPostings = ps
-
-	// If lazy expanded posting enabled, it is possible to fetch more series
-	//  so easier to hit the series limit.
-	if err := seriesLimiter.Reserve(uint64(len(ps.postings))); err != nil {
-		return httpgrpc.Errorf(int(codes.ResourceExhausted), "exceeded series limit: %s", err)
-	}
-
-	if b.batchSize > len(ps.postings) {
-		b.batchSize = len(ps.postings)
-	}
-	if b.lazyPostings.lazyExpanded() {
-		// Assume lazy expansion could cut actual expanded postings length to 50%.
-		b.expandedPostings = make([]storage.SeriesRef, 0, len(b.lazyPostings.postings)/2)
-		b.lazyExpandedPostingsCount.Inc()
-	}
-	b.entries = make([]seriesEntry, 0, b.batchSize)
-	return nil
-}
-
-func (b *blockSeriesClient) Recv() (*storepb.SeriesResponse, error) {
-	for len(b.entries) == 0 && b.hasMorePostings {
-		if err := b.nextBatch(b.tenant); err != nil {
-			return nil, err
-		}
-	}
-
-	if len(b.entries) == 0 {
-		b.seriesFetchDurationSum.WithLabelValues(b.tenant).Observe(b.indexr.stats.SeriesDownloadLatencySum.Seconds())
-		if b.chunkr != nil {
-			b.chunkFetchDuration.WithLabelValues(b.tenant).Observe(b.chunkr.stats.ChunksFetchDurationSum.Seconds())
-			b.chunkFetchDurationSum.WithLabelValues(b.tenant).Observe(b.chunkr.stats.ChunksDownloadLatencySum.Seconds())
-		}
-		return nil, io.EOF
-	}
-
-	next := b.entries[0]
-	b.entries = b.entries[1:]
-
-	return storepb.NewSeriesResponse(&storepb.Series{
-		Labels: labelpb.ZLabelsFromPromLabels(next.lset),
-		Chunks: next.chks,
-	}), nil
-}
-
-func (b *blockSeriesClient) nextBatch(tenant string) error {
-	start := b.i
-	end := start + uint64(b.batchSize)
-	if end > uint64(len(b.lazyPostings.postings)) {
-		end = uint64(len(b.lazyPostings.postings))
-	}
-	b.i = end
-
-	postingsBatch := b.lazyPostings.postings[start:end]
-	if len(postingsBatch) == 0 {
-		b.hasMorePostings = false
-		if b.lazyPostings.lazyExpanded() {
-			v, err := b.indexr.IndexVersion()
-			if err != nil {
-				return errors.Wrap(err, "get index version")
-			}
-			if v >= 2 {
-				for i := range b.expandedPostings {
-					b.expandedPostings[i] = b.expandedPostings[i] / 16
-				}
-			}
-			b.indexr.storeExpandedPostingsToCache(b.blockMatchers, index.NewListPostings(b.expandedPostings), len(b.expandedPostings), tenant)
-		}
-		return nil
-	}
-
-	b.indexr.reset(len(postingsBatch))
-	if !b.skipChunks {
-		b.chunkr.reset()
-	}
-
-	if err := b.indexr.PreloadSeries(b.ctx, postingsBatch, b.bytesLimiter, b.tenant); err != nil {
-		return errors.Wrap(err, "preload series")
-	}
-
-	b.entries = b.entries[:0]
-OUTER:
-	for i := 0; i < len(postingsBatch); i++ {
-		if err := b.ctx.Err(); err != nil {
-			return err
-		}
-		ok, err := b.indexr.LoadSeriesForTime(postingsBatch[i], &b.symbolizedLset, &b.chkMetas, b.skipChunks, b.mint, b.maxt)
-		if err != nil {
-			return errors.Wrap(err, "read series")
-		}
-		if !ok {
-			continue
-		}
-
-		if err := b.indexr.LookupLabelsSymbols(b.ctx, b.symbolizedLset, &b.lset); err != nil {
-			return errors.Wrap(err, "Lookup labels symbols")
-		}
-
-		for _, matcher := range b.lazyPostings.matchers {
-			val := b.lset.Get(matcher.Name)
-			if !matcher.Matches(val) {
-				// Series not matched means series we overfetched due to lazy posting expansion.
-				seriesBytes := b.indexr.loadedSeries[postingsBatch[i]]
-				b.lazyExpandedPostingSeriesOverfetchedSizeBytes.Add(float64(len(seriesBytes)))
-				continue OUTER
-			}
-		}
-		if b.lazyPostings.lazyExpanded() {
-			b.expandedPostings = append(b.expandedPostings, postingsBatch[i])
-		}
-
-		completeLabelset := labelpb.ExtendSortedLabels(b.lset, b.extLset)
-		if b.extLsetToRemove != nil {
-			completeLabelset = rmLabels(completeLabelset, b.extLsetToRemove)
-		}
-
-		if !b.shardMatcher.MatchesLabels(completeLabelset) {
-			continue
-		}
-
-		s := seriesEntry{lset: completeLabelset}
-		if b.skipChunks {
-			b.entries = append(b.entries, s)
-			continue
-		}
-
-		// Schedule loading chunks.
-		s.refs = make([]chunks.ChunkRef, 0, len(b.chkMetas))
-		s.chks = make([]storepb.AggrChunk, 0, len(b.chkMetas))
-
-		for j, meta := range b.chkMetas {
-			if err := b.chunkr.addLoad(meta.Ref, len(b.entries), j); err != nil {
-				return errors.Wrap(err, "add chunk load")
-			}
-			s.chks = append(s.chks, storepb.AggrChunk{
-				MinTime: meta.MinTime,
-				MaxTime: meta.MaxTime,
-			})
-			s.refs = append(s.refs, meta.Ref)
-		}
-
-		// Ensure sample limit through chunksLimiter if we return chunks.
-		if err := b.chunksLimiter.Reserve(uint64(len(b.chkMetas))); err != nil {
-			return httpgrpc.Errorf(int(codes.ResourceExhausted), "exceeded chunks limit: %s", err)
-		}
-
-		b.entries = append(b.entries, s)
-	}
-
-	if !b.skipChunks {
-		if err := b.chunkr.load(b.ctx, b.entries, b.loadAggregates, b.calculateChunkHash, b.bytesLimiter, b.tenant); err != nil {
-			return errors.Wrap(err, "load chunks")
-		}
-	}
-
-	return nil
 }
 
 func populateChunk(out *storepb.AggrChunk, in chunkenc.Chunk, aggrs []storepb.Aggr, save func([]byte) ([]byte, error), calculateChecksum bool) error {
@@ -3254,288 +2938,6 @@ func decodeSeriesForTime(b []byte, lset *[]symbolizedLabel, chks *[]chunks.Meta,
 		mint = maxt
 	}
 	return len(*chks) > 0, d.Err()
-}
-
-type loadIdx struct {
-	offset uint32
-	// Indices, not actual entries and chunks.
-	seriesEntry int
-	chunk       int
-}
-
-type bucketChunkReader struct {
-	block *bucketBlock
-
-	toLoad [][]loadIdx
-
-	// chunkBytesMtx protects access to chunkBytes, when updated from chunks-loading goroutines.
-	// After chunks are loaded, mutex is no longer used.
-	chunkBytesMtx sync.Mutex
-	stats         *queryStats
-	chunkBytes    []*[]byte // Byte slice to return to the chunk pool on close.
-
-	loadingChunksMtx  sync.Mutex
-	loadingChunks     bool
-	finishLoadingChks chan struct{}
-}
-
-func newBucketChunkReader(block *bucketBlock) *bucketChunkReader {
-	return &bucketChunkReader{
-		block:  block,
-		stats:  &queryStats{},
-		toLoad: make([][]loadIdx, len(block.chunkObjs)),
-	}
-}
-
-func (r *bucketChunkReader) reset() {
-	for i := range r.toLoad {
-		r.toLoad[i] = r.toLoad[i][:0]
-	}
-	r.loadingChunksMtx.Lock()
-	r.loadingChunks = false
-	r.finishLoadingChks = make(chan struct{})
-	r.loadingChunksMtx.Unlock()
-}
-
-func (r *bucketChunkReader) Close() error {
-	// NOTE(GiedriusS): we need to wait until loading chunks because loading
-	// chunks modifies r.block.chunkPool.
-	r.loadingChunksMtx.Lock()
-	loadingChks := r.loadingChunks
-	r.loadingChunksMtx.Unlock()
-
-	if loadingChks {
-		<-r.finishLoadingChks
-	}
-	r.block.pendingReaders.Done()
-
-	for _, b := range r.chunkBytes {
-		r.block.chunkPool.Put(b)
-	}
-	return nil
-}
-
-// addLoad adds the chunk with id to the data set to be fetched.
-// Chunk will be fetched and saved to refs[seriesEntry][chunk] upon r.load(refs, <...>) call.
-func (r *bucketChunkReader) addLoad(id chunks.ChunkRef, seriesEntry, chunk int) error {
-	var (
-		seq = int(id >> 32)
-		off = uint32(id)
-	)
-	if seq >= len(r.toLoad) {
-		return errors.Errorf("reference sequence %d out of range", seq)
-	}
-	r.toLoad[seq] = append(r.toLoad[seq], loadIdx{off, seriesEntry, chunk})
-	return nil
-}
-
-// load loads all added chunks and saves resulting aggrs to refs.
-func (r *bucketChunkReader) load(ctx context.Context, res []seriesEntry, aggrs []storepb.Aggr, calculateChunkChecksum bool, bytesLimiter BytesLimiter, tenant string) error {
-	r.loadingChunksMtx.Lock()
-	r.loadingChunks = true
-	r.loadingChunksMtx.Unlock()
-
-	begin := time.Now()
-	defer func() {
-		r.stats.ChunksDownloadLatencySum += time.Since(begin)
-
-		r.loadingChunksMtx.Lock()
-		r.loadingChunks = false
-		r.loadingChunksMtx.Unlock()
-
-		close(r.finishLoadingChks)
-	}()
-
-	g, ctx := errgroup.WithContext(ctx)
-
-	for seq, pIdxs := range r.toLoad {
-		sort.Slice(pIdxs, func(i, j int) bool {
-			return pIdxs[i].offset < pIdxs[j].offset
-		})
-		parts := r.block.partitioner.Partition(len(pIdxs), func(i int) (start, end uint64) {
-			return uint64(pIdxs[i].offset), uint64(pIdxs[i].offset) + uint64(r.block.estimatedMaxChunkSize)
-		})
-
-		for _, p := range parts {
-			if err := bytesLimiter.Reserve(uint64(p.End - p.Start)); err != nil {
-				return httpgrpc.Errorf(int(codes.ResourceExhausted), "exceeded bytes limit while fetching chunks: %s", err)
-			}
-			r.stats.DataDownloadedSizeSum += units.Base2Bytes(p.End - p.Start)
-		}
-
-		for _, p := range parts {
-			seq := seq
-			p := p
-			indices := pIdxs[p.ElemRng[0]:p.ElemRng[1]]
-			g.Go(func() error {
-				return r.loadChunks(ctx, res, aggrs, seq, p, indices, calculateChunkChecksum, bytesLimiter, tenant)
-			})
-		}
-	}
-	return g.Wait()
-}
-
-// loadChunks will read range [start, end] from the segment file with sequence number seq.
-// This data range covers chunks starting at supplied offsets.
-func (r *bucketChunkReader) loadChunks(ctx context.Context, res []seriesEntry, aggrs []storepb.Aggr, seq int, part Part, pIdxs []loadIdx, calculateChunkChecksum bool, bytesLimiter BytesLimiter, tenant string) error {
-	fetchBegin := time.Now()
-	stats := new(queryStats)
-	defer func() {
-		stats.ChunksFetchDurationSum += time.Since(fetchBegin)
-		r.stats.merge(stats)
-	}()
-
-	// Get a reader for the required range.
-	reader, err := r.block.chunkRangeReader(ctx, seq, int64(part.Start), int64(part.End-part.Start))
-	if err != nil {
-		return errors.Wrap(err, "get range reader")
-	}
-	defer runutil.CloseWithLogOnErr(r.block.logger, reader, "readChunkRange close range reader")
-	bufReader := bufio.NewReaderSize(reader, r.block.estimatedMaxChunkSize)
-
-	stats.chunksFetchCount++
-	stats.chunksFetched += len(pIdxs)
-	stats.ChunksFetchedSizeSum += units.Base2Bytes(int(part.End - part.Start))
-
-	var (
-		buf        []byte
-		readOffset = int(pIdxs[0].offset)
-
-		// Save a few allocations.
-		written  int
-		diff     uint32
-		chunkLen int
-		n        int
-	)
-
-	bufPooled, err := r.block.chunkPool.Get(r.block.estimatedMaxChunkSize)
-	if err == nil {
-		buf = *bufPooled
-	} else {
-		buf = make([]byte, r.block.estimatedMaxChunkSize)
-	}
-	defer r.block.chunkPool.Put(&buf)
-
-	for i, pIdx := range pIdxs {
-		// Fast forward range reader to the next chunk start in case of sparse (for our purposes) byte range.
-		for readOffset < int(pIdx.offset) {
-			written, err = bufReader.Discard(int(pIdx.offset) - int(readOffset))
-			if err != nil {
-				return errors.Wrap(err, "fast forward range reader")
-			}
-			readOffset += written
-		}
-		// Presume chunk length to be reasonably large for common use cases.
-		// However, declaration for EstimatedMaxChunkSize warns us some chunks could be larger in some rare cases.
-		// This is handled further down below.
-		chunkLen = r.block.estimatedMaxChunkSize
-		if i+1 < len(pIdxs) {
-			if diff = pIdxs[i+1].offset - pIdx.offset; int(diff) < chunkLen {
-				chunkLen = int(diff)
-			}
-		}
-		cb := buf[:chunkLen]
-		n, err = io.ReadFull(bufReader, cb)
-		readOffset += n
-		// Unexpected EOF for last chunk could be a valid case. Any other errors are definitely real.
-		if err != nil && !(errors.Is(err, io.ErrUnexpectedEOF) && i == len(pIdxs)-1) {
-			return errors.Wrapf(err, "read range for seq %d offset %x", seq, pIdx.offset)
-		}
-
-		chunkDataLen, n := binary.Uvarint(cb)
-		if n < 1 {
-			return errors.New("reading chunk length failed")
-		}
-
-		// Chunk length is n (number of bytes used to encode chunk data), 1 for chunk encoding and chunkDataLen for actual chunk data.
-		// There is also crc32 after the chunk, but we ignore that.
-		chunkLen = n + 1 + int(chunkDataLen)
-		if chunkLen <= len(cb) {
-			err = populateChunk(&(res[pIdx.seriesEntry].chks[pIdx.chunk]), rawChunk(cb[n:chunkLen]), aggrs, r.save, calculateChunkChecksum)
-			if err != nil {
-				return errors.Wrap(err, "populate chunk")
-			}
-			stats.chunksTouched++
-			stats.ChunksTouchedSizeSum += units.Base2Bytes(int(chunkDataLen))
-			continue
-		}
-
-		r.block.metrics.chunkRefetches.WithLabelValues(tenant).Inc()
-		// If we didn't fetch enough data for the chunk, fetch more.
-		fetchBegin = time.Now()
-		// Read entire chunk into new buffer.
-		// TODO: readChunkRange call could be avoided for any chunk but last in this particular part.
-		if err := bytesLimiter.Reserve(uint64(chunkLen)); err != nil {
-			return httpgrpc.Errorf(int(codes.ResourceExhausted), "exceeded bytes limit while fetching chunks: %s", err)
-		}
-		stats.DataDownloadedSizeSum += units.Base2Bytes(chunkLen)
-
-		nb, err := r.block.readChunkRange(ctx, seq, int64(pIdx.offset), int64(chunkLen), []byteRange{{offset: 0, length: chunkLen}})
-		if err != nil {
-			return errors.Wrapf(err, "preloaded chunk too small, expecting %d, and failed to fetch full chunk", chunkLen)
-		}
-		if len(*nb) != chunkLen {
-			return errors.Errorf("preloaded chunk too small, expecting %d", chunkLen)
-		}
-
-		stats.chunksFetchCount++
-		stats.ChunksFetchedSizeSum += units.Base2Bytes(len(*nb))
-		err = populateChunk(&(res[pIdx.seriesEntry].chks[pIdx.chunk]), rawChunk((*nb)[n:]), aggrs, r.save, calculateChunkChecksum)
-		if err != nil {
-			r.block.chunkPool.Put(nb)
-			return errors.Wrap(err, "populate chunk")
-		}
-		stats.chunksTouched++
-		stats.ChunksTouchedSizeSum += units.Base2Bytes(int(chunkDataLen))
-
-		r.block.chunkPool.Put(nb)
-	}
-	return nil
-}
-
-// save saves a copy of b's payload to a memory pool of its own and returns a new byte slice referencing said copy.
-// Returned slice becomes invalid once r.block.chunkPool.Put() is called.
-func (r *bucketChunkReader) save(b []byte) ([]byte, error) {
-	r.chunkBytesMtx.Lock()
-	defer r.chunkBytesMtx.Unlock()
-	// Ensure we never grow slab beyond original capacity.
-	if len(r.chunkBytes) == 0 ||
-		cap(*r.chunkBytes[len(r.chunkBytes)-1])-len(*r.chunkBytes[len(r.chunkBytes)-1]) < len(b) {
-		s, err := r.block.chunkPool.Get(len(b))
-		if err != nil {
-			return nil, errors.Wrap(err, "allocate chunk bytes")
-		}
-		r.chunkBytes = append(r.chunkBytes, s)
-	}
-	slab := r.chunkBytes[len(r.chunkBytes)-1]
-	*slab = append(*slab, b...)
-	return (*slab)[len(*slab)-len(b):], nil
-}
-
-// rawChunk is a helper type that wraps a chunk's raw bytes and implements the chunkenc.Chunk
-// interface over it.
-// It is used to Store API responses which don't need to introspect and validate the chunk's contents.
-type rawChunk []byte
-
-func (b rawChunk) Encoding() chunkenc.Encoding {
-	return chunkenc.Encoding(b[0])
-}
-
-func (b rawChunk) Bytes() []byte {
-	return b[1:]
-}
-func (b rawChunk) Compact() {}
-
-func (b rawChunk) Iterator(_ chunkenc.Iterator) chunkenc.Iterator {
-	panic("invalid call")
-}
-
-func (b rawChunk) Appender() (chunkenc.Appender, error) {
-	panic("invalid call")
-}
-
-func (b rawChunk) NumSamples() int {
-	panic("invalid call")
 }
 
 type queryStats struct {
