+++ pkg/storegateway/storepb/rpc.proto
@@ -226,3 +226,83 @@ message LabelValuesResponse {
     /// implementation of a specific store.
     google.protobuf.Any hints = 3;
 }
+
+/// IndexStore represents API against instance that stores XOR encoded values with label set metadata (e.g Prometheus metrics).
+service IndexStore {
+    /// Series streams each Series (Labels and chunk/downsampling chunk) for given label matchers and time range.
+    rpc Select(SelectRequest) returns (stream SelectResponse);
+
+    /// LabelNames returns all label names constrained by the given matchers.
+    rpc LabelNames(LabelNamesRequest) returns (LabelNamesResponse);
+
+    /// LabelValues returns all label values for given label name.
+    rpc LabelValues(LabelValuesRequest) returns (LabelValuesResponse);
+}
+
+message SelectRequest {
+    int64 min_time = 1;
+    int64 max_time = 2;
+    repeated LabelMatcher matchers = 3 [(gogoproto.nullable) = false];
+
+    int64 max_resolution_window = 4;
+    repeated Aggr aggregates = 5;
+
+    // Deprecated. Use partial_response_strategy instead.
+    bool partial_response_disabled = 6;
+
+    // TODO(bwplotka): Move Thanos components to use strategy instead. Including QueryAPI.
+    PartialResponseStrategy partial_response_strategy = 7;
+
+    // skip_chunks controls whether sending chunks or not in series responses.
+    bool skip_chunks = 8;
+
+    // hints is an opaque data structure that can be used to carry additional information.
+    // The content of this field and whether it's supported depends on the
+    // implementation of a specific store.
+    google.protobuf.Any hints = 9;
+
+    // Query step size in milliseconds.
+    // Deprecated: Use query_hints instead.
+    int64 step = 10;
+
+    // Range vector selector range in milliseconds.
+    // Deprecated: Use query_hints instead.
+    int64 range = 11;
+
+    // query_hints are the hints coming from the PromQL engine when
+    // requesting a storage.SeriesSet for a given expression.
+    // As hints name suggest using those is best effort.
+    QueryHints query_hints = 12;
+
+    // shard_info is used by the querier to request a specific
+    // shard of blocks instead of entire blocks.
+    ShardInfo shard_info = 13;
+
+    // without_replica_labels are replica labels which have to be excluded from series set results.
+    // The sorting requirement has to be preserved, so series should be sorted without those labels.
+    // If the requested label is NOT a replica label (labels that identify replication group) it should be not affected by
+    // this setting (label should be included in sorting and response).
+    // It is the server responsibility to detect and track what is replica label and what is not.
+    // This allows faster deduplication by clients.
+    // NOTE(bwplotka): thanos.info.store.supports_without_replica_labels field has to return true to let client knows
+    // server supports it.
+    repeated string without_replica_labels = 14;
+}
+
+message SelectResponse {
+    oneof result {
+        /// series contains 1 response series. The series labels are sorted by name.
+        SelectedSeries series = 1;
+
+        /// warning is considered an information piece in place of series for warning purposes.
+        /// It is used to warn store API user about suspicious cases or partial response (if enabled).
+        string warning = 2;
+
+        /// hints is an opaque data structure that can be used to carry additional information from
+        /// the store. The content of this field and whether it's supported depends on the
+        /// implementation of a specific store. It's also implementation specific if it's allowed that
+        /// multiple SeriesResponse frames contain hints for a single Series() request and how should they
+        /// be handled in such case (ie. merged vs keep the first/last one).
+        google.protobuf.Any hints = 3;
+    }
+}
